backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.4.0
imagehash==4.3.1
numpy==2.1.1
web3==6.20.1
requests==2.32.3
